chore(build): don't try to deploy on actual pr

+++ devdeploy.sh
@@ -1,25 +1,29 @@
 SSHSRV="travis@fgpv.cloudapp.net"
 DESTDIR="/home/travis/www"
 
-openssl aes-256-cbc -k "$PW" -out ~/.ssh/id_rsa -in devkey.enc -d
-echo -e "Host *\n\tStrictHostKeyChecking no\n" >> ~/.ssh/config
-chmod 600 ~/.ssh/id_rsa
-eval `ssh-agent -s`
-ssh-add ~/.ssh/id_rsa
+if [ "$TRAVIS_PULL_REQUEST" == "false" ]; then
 
-if [ "$TRAVIS_REPO_SLUG" == "fgpv-vpgf/fgpv-vpgf" ]; then
+    openssl aes-256-cbc -k "$PW" -out ~/.ssh/id_rsa -in devkey.enc -d
+    echo -e "Host *\n\tStrictHostKeyChecking no\n" >> ~/.ssh/config
+    chmod 600 ~/.ssh/id_rsa
+    eval `ssh-agent -s`
+    ssh-add ~/.ssh/id_rsa
 
-    if [ -n "$TRAVIS_TAG" ]; then
-        DESTDIR="$DESTDIR/$TRAVIS_TAG/"
+    if [ "$TRAVIS_REPO_SLUG" == "fgpv-vpgf/fgpv-vpgf" ]; then
+
+        if [ -n "$TRAVIS_TAG" ]; then
+            DESTDIR="$DESTDIR/$TRAVIS_TAG/"
+        else
+            DESTDIR="$DESTDIR/$TRAVIS_BRANCH/"
+        fi
+
     else
-        DESTDIR="$DESTDIR/$TRAVIS_BRANCH/"
+        USER=${TRAVIS_REPO_SLUG/\/fgpv-vpgf/}
+        DESTDIR="$DESTDIR/users/$USER/$TRAVIS_BRANCH/"
     fi
 
-else
-    USER=${TRAVIS_REPO_SLUG/\/fgpv-vpgf/}
-    DESTDIR="$DESTDIR/users/$USER/$TRAVIS_BRANCH/"
-fi
+    echo "Destintation: $DESTDIR"
+    ssh "$SSHSRV" mkdir -p "$DESTDIR"
+    rsync -av --delete "build/" "$SSHSRV:$DESTDIR"
 
-echo "Destintation: $DESTDIR"
-ssh "$SSHSRV" mkdir -p "$DESTDIR"
-rsync -av --delete "build/" "$SSHSRV:$DESTDIR"
+fi
